feat(pt): move po;itic-fact-check to shared component and add isLanding property
